backend/requirements-prod.txt
# Core Flask dependencies - essential for the app to run
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0

# Database and vector storage
# ChromaDB with all required dependencies
chromadb==0.4.15
# Core ChromaDB dependencies (pinned to avoid NumPy 2.0 compatibility issues)
numpy==1.24.4
sqlalchemy>=1.4.0,<2.0.0
pydantic>=1.8.0,<2.0.0
typing-extensions>=4.0.0
# Additional ChromaDB system dependencies
psutil>=5.8.0
# ChromaDB embedding dependencies
sentence-transformers>=2.2.0
# ChromaDB vector store dependencies
hnswlib>=0.7.0
# ChromaDB HTTP client dependencies
httpx>=0.24.0
# ChromaDB persistence dependencies
duckdb>=0.8.0

# Auth and security (lightweight subset)
PyJWT==2.9.0
bcrypt==4.0.1
cffi==1.16.0
# Additional auth dependencies
pyjwt==2.9.0

# Email (optional)
flask-mail==0.9.1
itsdangerous==2.1.2
werkzeug==2.3.7

# Production notes:
# - sentence-transformers and torch are excluded to reduce build time and image size
# - These can be added later if AI features are needed in production
# - Consider using a separate AI service or lighter alternatives for production
